Order api

+++ packages/api/src/utils/assets/removeUpload.ts
@@ -2,8 +2,11 @@ import fs from 'fs';
 
 export async function removeUpload(path: string) {
   try {
-    fs.unlinkSync(path);
+    if (fs.existsSync(path)) {
+      fs.unlinkSync(path);
+    }
   } catch (e) {
+    console.log(`Error in removeUpload`);
     console.log(e);
   }
 }
